first time we can record real data
svn-id: https://kwave.svn.sourceforge.net/svnroot/kwave/trunk@1137 bbf32e5a-961d-0410-9e10-91f32741c31e

+++ plugins/record/RecordThread.cpp
@@ -156,7 +156,7 @@ void RecordThread::run()
 	while (len) {
 	    // read raw data from the record device
 	    int result = m_device->read(p, len);
-	    Q_ASSERT(result == len);
+	    Q_ASSERT(result == (int)len);
 
 	    if (result == -EAGAIN) {
 		// thread was interrupted, received signal?
